Added some tests.

+++ compiler/views.py
@@ -1,3 +1,4 @@
+from django.contrib.auth.decorators import login_required
 from django.http import JsonResponse
 from django.shortcuts import get_object_or_404, render
 from django.views.decorators.csrf import csrf_exempt
@@ -17,39 +18,45 @@ def index(request):
     return render(request, 'compiler/index.html', context)
 
 
+@login_required
 def compile_file(request, file_info_id):
-    if request.method == 'POST':
+    if request.method == 'POST' and utils.test_ownership(request.user, file_info_id):
         context = dict()
         # Get file
         file_info = get_object_or_404(FileInfo, pk=file_info_id)
         file = get_object_or_404(File, info=file_info)
         # Get compilation options
         form = CompileForm(request.POST)
-        options = utils.get_options(form)
-        # Compiling the file, saving the output to the temporary directory
-        import os
-        temporary_path = file.content.path.replace("media/files", "temporary")
-        os.chdir(os.path.dirname(temporary_path))
-        # Compile with options
-        cmd = 'sdcc' + ' -S ' + options + ' ' + file.content.path
-        context['cmd'] = cmd
-        err_message = os.popen(cmd + ' 2>&1').read()
-        status_code = 200
-        # Get output
-        try:
-            with open(temporary_path.replace('.c', '.asm'), 'r') as output_file:
-                [headers, bodies] = utils.separate_assembly_sections(output_file.read())
-                context['headers'] = headers
-                context['bodies'] = bodies
-        except FileNotFoundError:
-            context['errors'] = utils.parse_err_message(err_message)
-            status_code = 400
-        # Remove all files from temporary directory
-        os.system('rm -rf ' + os.path.dirname(temporary_path) + '/*')
-        return JsonResponse(status=status_code, data=context)
+        if form.is_valid():
+            options = utils.get_options(form)
+            # Compiling the file, saving the output to the temporary directory
+            import os
+            temporary_path = file.content.path.replace("media/files", "temporary")
+            os.chdir(os.path.dirname(temporary_path))
+            # Compile with options
+            cmd = 'sdcc' + ' -S ' + options + ' ' + file.content.path
+            context['cmd'] = cmd
+            err_message = os.popen(cmd + ' 2>&1').read()
+            status_code = 200
+            # Get output
+            try:
+                with open(temporary_path.replace('.c', '.asm'), 'r') as output_file:
+                    [headers, bodies] = utils.separate_assembly_sections(output_file.read())
+                    context['headers'] = headers
+                    context['bodies'] = bodies
+            except FileNotFoundError:
+                context['errors'] = utils.parse_err_message(err_message)
+                status_code = 400
+            # Remove all files from temporary directory
+            os.system('rm -rf ' + os.path.dirname(temporary_path) + '/*')
+            return JsonResponse(status=status_code, data=context)
+    return utils.JSON_BAD_REQUEST
 
 
+@login_required
 def show_file(request, file_info_id):
+    if not utils.test_ownership(request.user, file_info_id):
+        return utils.JSON_BAD_REQUEST
     file_info = get_object_or_404(FileInfo, pk=file_info_id)
     file = get_object_or_404(File, info=file_info)
     content = file.content.read().decode('utf-8')
@@ -57,12 +64,11 @@ def show_file(request, file_info_id):
     return JsonResponse({'code': code})
 
 
+@login_required
 def change_sections(request, file_info_id):
-    if request.method == 'POST':
+    if request.method == 'POST' and utils.test_ownership(request.user, file_info_id):
         form = ChangeSectionsForm(request.POST)
         if form.is_valid():
-            context = dict()
-            # Get file
             file_info = get_object_or_404(FileInfo, pk=file_info_id)
             file = get_object_or_404(File, info=file_info)
             # Get sections
@@ -70,17 +76,17 @@ def change_sections(request, file_info_id):
                                    form.cleaned_data['end'],
                                    form.cleaned_data['sectionType'],
                                    file)
-            # Null response
-            return JsonResponse(context)
+            return utils.JSON_OK
         else:
             print(form.errors)
+    return utils.JSON_BAD_REQUEST
 
 
+@login_required
 def upload_file(request):
     if request.method == 'POST':
         form = UploadFileForm(request.POST, request.FILES)
         if form.is_valid():
-            context = dict()
             # Parse form
             name = form.cleaned_data['file'].name
             description = form.cleaned_data['description']
@@ -93,17 +99,17 @@ def upload_file(request):
             file = File.objects.create(info=file_info, parent=parent, content=form.cleaned_data['file'])
             utils.create_file_sections(file.content.read().decode('utf-8'), file)
             file.save()
-            # Null response
-            return JsonResponse(context)
+            return utils.JSON_OK
         else:
             print(form.errors)
+    return utils.JSON_BAD_REQUEST
 
 
+@login_required
 def create_directory(request):
     if request.method == 'POST':
         form = CreateDirectoryForm(request.POST)
         if form.is_valid():
-            context = dict()
             # Parse form
             name = form.cleaned_data['name']
             description = form.cleaned_data['description']
@@ -114,22 +120,22 @@ def create_directory(request):
             directory_info = FileInfo.objects.create(name=name, description=description, owner=request.user)
             directory = Directory.objects.create(info=directory_info, parent=parent)
             directory.save()
-            # Null response
-            return JsonResponse(context)
+            return utils.JSON_OK
         else:
             print(form.errors)
+    return utils.JSON_BAD_REQUEST
 
 
+@login_required
 @csrf_exempt
 def delete_file(request, file_info_id):
-    if request.method == 'DELETE':
-        context = dict()
+    if request.method == 'DELETE' and utils.test_ownership(request.user, file_info_id):
         # Get file
         file_info = get_object_or_404(FileInfo, pk=file_info_id)
         file_info.available = False
         file_info.save()
-        # Null response
-        return JsonResponse(context)
+        return utils.JSON_OK
+    return utils.JSON_BAD_REQUEST
 
 
 def directories_tree(request):
